additional fix for coverity 29981 (SCT_Monitoring-00-06-53)
Former-commit-id: c51dfc5d35314b93c62b025fa6a6078dfd845ea7

+++ InnerDetector/InDetMonitoring/SCT_Monitoring/SCT_Monitoring/SCTTracksMonTool.h
@@ -178,7 +178,7 @@ private:
   //@name Service methods
   //@{
   /// Calculate Pull value for MeasuredAtPlane TrackStates
-  float calculatePull(const float, const float, const float);
+  double calculatePull(const double, const double, const double);
   
   ///Factory + register for the 2D histos, returns whether successfully registered
   StatusCode h2Factory(const std::string & name, const std::string & title, const SCT_Monitoring::Bec bec, MonGroup & registry, VecH2_t & storageVector);
